requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
python-dotenv==1.0.0
fpdf2==2.7.4
pillow==10.0.0
gunicorn==21.2.0
wtforms==3.0.1